day 11 mods
Switched to a system of keeping track of the number of unsettled seats.
If a seat has not enough seats around it to ever make it unoccupied, use
that knowledge. Likewise, if a seat has more than num seats around it
that are already settled, mark it.

This makes the algo run in about half the time :shrug: